// src/proto/steam/steam.ts
export interface InventoryItem {
    /**
     * @generated from protobuf field: string asset_id = 1
     */
    assetId: string;
    /**
     * @generated from protobuf field: string class_id = 2
     */
    classId: string;
    /**
     * @generated from protobuf field: string name = 3
     */
    name: string;
    /**
     * @generated from protobuf field: bool tradable = 4
     */
    tradable: boolean;
    /**
     * @generated from protobuf field: string icon_url = 5
     */
    iconUrl: string;
    /**
     * @generated from protobuf field: string quality = 6
     */
    quality: string;
    /**
     * @generated from protobuf field: string rarity = 7
     */
    rarity: string;
    /**
     * @generated from protobuf field: bool is_case = 8
     */
    isCase: boolean; // Validated against whitelist
}
// ...
class InventoryItem$Type extends MessageType<InventoryItem> {
    constructor() {
        super("steam.InventoryItem", [
            { no: 1, name: "asset_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "class_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "tradable", kind: "scalar", T: 8 /*ScalarType.BOOL*/ },
            { no: 5, name: "icon_url", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "quality", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 7, name: "rarity", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 8, name: "is_case", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<InventoryItem>): InventoryItem {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.assetId = "";
        message.classId = "";
        message.name = "";
        message.tradable = false;
        message.iconUrl = "";
        message.quality = "";
        message.rarity = "";
        message.isCase = false;
        if (value !== undefined)
            reflectionMergePartial<InventoryItem>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: InventoryItem): InventoryItem {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string asset_id */ 1:
                    message.assetId = reader.string();
                    break;
                case /* string class_id */ 2:
                    message.classId = reader.string();
                    break;
                case /* string name */ 3:
                    message.name = reader.string();
                    break;
                case /* bool tradable */ 4:
                    message.tradable = reader.bool();
                    break;
                case /* string icon_url */ 5:
                    message.iconUrl = reader.string();
                    break;
                case /* string quality */ 6:
                    message.quality = reader.string();
                    break;
                case /* string rarity */ 7:
                    message.rarity = reader.string();
                    break;
                case /* bool is_case */ 8:
                    message.isCase = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: InventoryItem, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string asset_id = 1; */
        if (message.assetId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.assetId);
        /* string class_id = 2; */
        if (message.classId !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.classId);
        /* string name = 3; */
        if (message.name !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.name);
        /* bool tradable = 4; */
        if (message.tradable !== false)
            writer.tag(4, WireType.Varint).bool(message.tradable);
        /* string icon_url = 5; */
        if (message.iconUrl !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.iconUrl);
        /* string quality = 6; */
        if (message.quality !== "")
            writer.tag(6, WireType.LengthDelimited).string(message.quality);
        /* string rarity = 7; */
        if (message.rarity !== "")
            writer.tag(7, WireType.LengthDelimited).string(message.rarity);
        /* bool is_case = 8; */
        if (message.isCase !== false)
            writer.tag(8, WireType.Varint).bool(message.isCase);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message steam.InventoryItem
 */
export const InventoryItem = new InventoryItem$Type();